rezvert LevelNumber

+++ src/globals.h
@@ -255,7 +255,7 @@ typedef int16_t CrAttackType;
 /** Creature death flags, stores CrDed_* values. */
 typedef uint16_t CrDeathFlags;
 /** Level number within a campaign. */
-typedef int32_t LevelNumber;
+typedef long LevelNumber;
 /** Game turn number, used for in-game time computations. */
 typedef uint32_t GameTurn;
 /** Game turns difference, used for in-game time computations. */
